Move the options to their own implementation file
The options for controling the hasher are not part of the api event
schema. Putting them in their own file makes it clear their use is the
responsibility of the caller rather than formally part of the event
schema.
